Board Backend
Contains the current progress of the board in its purely backend state. Unplayable spots are represented with a -1, and empty playable spots with 0's. I've included the ability to check adjacencies for a mill in the check_adjacent function, which should only be used after a piece has been placed to check for a new mill in that row or column of adjacent spots.