Mill Removal Update
Added the functionality so that when mills are found that valid pieces can be removed. While it is currently functional in this state, my algorithm to check for valid removal spots (pieces that are not in a mill unless all opposing pieces are in a mill) does not account for scenarios of all of their pieces being in 2 or 3 mills, only 1. We might have to edit code, or just create a function that checks to see if all of the opponent's pieces are in mills. Probably the latter would be easier, but I'm not 100% sure.

Other than that, I cleaned up some small bugs I saw, and edited some lines to get rid of weak warnings from my PyCharm IDE's warning system.