Player Class Refactor
Added in the player class and refactored the code to work with that. This results in a lot less global variables and a more streamlined system. Also changed the class from game_functons to game_functions lol, idk how we didn't notice that. Made some minor changes to test_functions as well as main to reflect the changes made elsewhere, and I'll note that I'm not 100% on the test_functions class, but it should be all okay. Besides the test_functions class though, everything has been tested and should work all the same from a user perspective as it did before.