Basic Win Condition Update
Looked through the updates made, played around with all of the new features and everything looked great! Only made a small update that checks for win condition and tells the user(s) when said win condition has occurred. I want to edit this more in the coming days, but I won't be able to much for the next few days and wanted to get this version out for Daylan so that there was no repetition of work. 
Big side note! I've found a bug that I'm not entirely sure how to patch out right now, so I'm posting a link to what happens during this bad edge case for now in case anyone wants to take a swing at it while I can't help out. If a mill is formed when a player places their last piece during phase 1, the program will show all adjacent squares in orange, even though it is a time to remove a piece from the opponent. The piece(s) highlighted in orange will stay highlighted until a player on a different turn is ready to legitimately move a piece. Here's a link to an imgur of a picture of what I am talking about in case this all was not clear. https://imgur.com/gallery/1tP544b

As always with me, just let me let me know if you have any questions, and I'll always be happy to respond whenever I can.